# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "An application that implements the basic features of stack overflow",
    "version": "1.0.0",
    "title": "Softcom Take Home Test",
    "contact": {
      "email": "mogbeyidavid@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-auth-token",
      "in": "header"
    }
  },
  "paths": {
...
    "/users/reset-password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Reset a user's password",
        "operationId": "reset-password",
        "description": "Reset a user's password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "Password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUserpassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/questions": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Ask a new question",
        "operationId": "create-question",
        "description": "Ask a new question",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "api_key": [
              "x-auth-token"
            ]
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "question",
            "description": "new question to be asked",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Question"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get all questions",
        "operationId": "get-questions",
        "description": "Get all questions",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/answers/{id}": {
      "get": {
        "tags": [
          "Answers"
        ],
        "summary": "Get one answer",
        "operationId": "get-answer",
        "description": "Get an answer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Answers"
        ],
        "summary": "update an answer",
        "operationId": "update-answer",
        "description": "Update a answer",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "api_key": [
              "x-auth-token"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "update an answer",
            "description": "Update an answer",
            "schema": {
              "$ref": "#/definitions/UpdateAnswer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "NewUserpassword": {
      "type": "object",
      "required": [
        "password",
        "confirmPassword"
      ],
      "properties": {
        "password": {
          "type": "string",
          "format": "password"
        },
        "confirmPassword": {
          "type": "string",
          "format": "password"
        }
      }
    },
    "Question": {
      "type": "object",
      "required": [
        "question"
      ],
      "properties": {
        "question": {
          "type": "string"
        }
      }
    },
...
    "UpdateAnswer": {
      "type": "object",
      "required": [
        "answer"
      ],
      "properties": {
        "question": {
          "type": "string"
        }
      }
    }
  },
  "schemes": [
    "http"
  ],
  "basePath": "/api/v1",
  "host": "127.0.0.1"
}